[TOY-44] org.contribhub.core.service -> org.contribhub.core rename
