+++ api/src/main/kotlin/org/contribhub/api/common/response/ResponseService.kt
@@ -1,7 +1,7 @@
 package org.contribhub.api.common.response
 
 import org.contribhub.api.common.response.exception.CustomExceptionResponse
-import org.contribhub.core.service.exception.CustomExceptionStatus
+import org.contribhub.core.exception.CustomExceptionStatus
 import org.contribhub.api.common.response.success.CustomSuccessResponse
 import org.contribhub.api.common.response.success.CustomSuccessStatus
 import org.springframework.stereotype.Service
