+++ api/src/main/kotlin/org/contribhub/api/controller/filter/GithubAuthFilter.kt
@@ -4,8 +4,8 @@ import jakarta.servlet.FilterChain
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
 import kotlinx.coroutines.runBlocking
-import org.contribhub.core.service.entity.UserInfo
-import org.contribhub.core.service.service.UserAuthService
+import org.contribhub.core.entity.UserInfo
+import org.contribhub.core.service.UserAuthService
 import org.springframework.http.HttpHeaders
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
 import org.springframework.security.core.context.SecurityContextHolder
