+++ api/src/main/kotlin/org/contribhub/api/dto/response/RepositoryDetailResponse.kt
@@ -1,6 +1,6 @@
 package org.contribhub.api.dto.response
 
-import org.contribhub.core.service.entity.RepositoryDetail
+import org.contribhub.core.entity.RepositoryDetail
 
 data class RepositoryDetailResponse(
     // contribhub의 레포지토리 테이블 시퀀스
