+++ api/src/main/kotlin/org/contribhub/api/dto/searchkeyword/LicenseResponse.kt
@@ -2,7 +2,7 @@ package org.contribhub.api.dto.searchkeyword
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies
 import com.fasterxml.jackson.databind.annotation.JsonNaming
-import org.contribhub.core.service.entity.License
+import org.contribhub.core.entity.License
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy::class)
 data class LicenseResponse(
