+++ core/src/main/kotlin/org/contribhub/core/entity/License.kt
@@ -1,4 +1,4 @@
-package org.contribhub.core.service.entity
+package org.contribhub.core.entity
 
 data class License(
     val name: String
