+++ core/src/main/kotlin/org/contribhub/core/entity/RepositorySearchKey.kt
@@ -1,4 +1,4 @@
-package org.contribhub.core.service.entity
+package org.contribhub.core.entity
 
 // todo domain에 있기에는 살짝 어색하다는 느낌
 data class RepositorySearchKey(
