+++ core/src/main/kotlin/org/contribhub/core/entity/Topic.kt
@@ -1,0 +1,5 @@
+package org.contribhub.core.entity
+
+data class Topic(
+    val name: String
+)
