+++ core/src/main/kotlin/org/contribhub/core/entity/UserInfo.kt
@@ -1,4 +1,4 @@
-package org.contribhub.core.service.entity
+package org.contribhub.core.entity
 
 data class UserInfo(
     val id: Long,
