+++ core/src/main/kotlin/org/contribhub/core/exception/CustomException.kt
@@ -1,4 +1,4 @@
-package org.contribhub.core.service.exception
+package org.contribhub.core.exception
 
 data class CustomException(
     val customExceptionStatus: CustomExceptionStatus,
