+++ core/src/main/kotlin/org/contribhub/core/repository/GithubAuthRepository.kt
@@ -1,6 +1,6 @@
-package org.contribhub.core.service.repository
+package org.contribhub.core.repository
 
-import org.contribhub.core.service.entity.UserInfo
+import org.contribhub.core.entity.UserInfo
 
 interface GithubAuthRepository {
     suspend fun getAuthenticatedUser(accessToken: String): UserInfo
