--- core/src/main/kotlin/org/contribhub/core/service/entity/Topic.kt
@@ -1,5 +1,0 @@
-package org.contribhub.core.service.entity
-
-data class Topic(
-    val name: String
-)
