+++ infra/src/main/kotlin/org/contribhub/infra/http/dto/TopicInfoDTO.kt
@@ -1,6 +1,6 @@
 package org.contribhub.infra.http.dto
 
-import org.contribhub.core.service.entity.Topic
+import org.contribhub.core.entity.Topic
 
 data class TopicInfoDTO(
     val name: String,
