+++ infra/src/main/kotlin/org/contribhub/infra/repository/entity/LanguageEntity.kt
@@ -7,7 +7,7 @@ import jakarta.persistence.GenerationType
 import jakarta.persistence.Id
 import jakarta.persistence.Table
 import jakarta.persistence.UniqueConstraint
-import org.contribhub.core.service.entity.Language
+import org.contribhub.core.entity.Language
 import org.contribhub.infra.repository.entity.common.BaseTimeEntity
 
 @Entity
