+++ infra/src/main/kotlin/org/contribhub/infra/repository/entity/LicenseEntity.kt
@@ -7,7 +7,7 @@ import jakarta.persistence.GenerationType
 import jakarta.persistence.Id
 import jakarta.persistence.Table
 import jakarta.persistence.UniqueConstraint
-import org.contribhub.core.service.entity.License
+import org.contribhub.core.entity.License
 import org.contribhub.infra.repository.entity.common.BaseTimeEntity
 
 @Entity
