+++ infra/src/main/kotlin/org/contribhub/infra/repository/entity/TopicEntity.kt
@@ -10,7 +10,7 @@ import jakarta.persistence.JoinColumn
 import jakarta.persistence.ManyToOne
 import jakarta.persistence.Table
 import jakarta.persistence.UniqueConstraint
-import org.contribhub.core.service.entity.Topic
+import org.contribhub.core.entity.Topic
 import org.contribhub.infra.repository.entity.common.BaseTimeEntity
 
 @Entity
